backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
openai==1.12.0
numpy==1.26.4
scikit-learn==1.4.1
nltk==3.8.1
python-dotenv==1.0.1
gunicorn==21.2.0